# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AssistantQuestionInput {
  /**
     * @minLength 1
     * @maxLength 500
     */
  question: string;
}

export interface AssistantAnswer {
  answer: string;
  sources: string[];
}