# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: classiq
    description: ClassIQ live-data assistant
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /classiq/assistant:
    post:
      operationId: askClassroomAssistant
      tags: [classiq]
      summary: Ask a question about live classroom data
      description: Returns a deterministic answer derived from the authenticated user's current Supabase data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantQuestionInput"
      responses:
        "200":
          description: Answer grounded in current classroom data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantAnswer"
        "401":
          description: Authentication required
        "503":
          description: Supabase is not configured or unavailable
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AssistantQuestionInput:
      type: object
      properties:
        question:
          type: string
          minLength: 1
          maxLength: 500
      required:
        - question
    AssistantAnswer:
      type: object
      properties:
        answer:
          type: string
        sources:
          type: array
          items:
            type: string
      required:
        - answer
        - sources