# lib/api-zod/src/generated/types/assistantQuestionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AssistantQuestionInput {
  /**
     * @minLength 1
     * @maxLength 500
     */
  question: string;
}